requirements.txt
prettytable
networkx
pyvis
openai
requests
tenacity
python-dotenv